Feature/manage customer booking requests

+++ src/main/java/com/g42/platform/gms/auth/entity/StaffPrincipal.java
@@ -16,7 +16,7 @@ public class StaffPrincipal implements UserDetails {
         this.staffAuth = staffAuth;
     }
 
-    public Long getAuthId() {
+    public Integer getAuthId() {
         return staffAuth.getStaffAuthId();
     }
